wip: feat(amm): complete stable pool and middle price computation, add file structure
