+++ src/parachain/AMM.ts
@@ -2,274 +2,14 @@ import { MonetaryAmount } from "@interlay/monetary-js";
 import { ApiPromise } from "@polkadot/api";
 import { AddressOrPair } from "@polkadot/api/types";
 import { AccountId } from "@polkadot/types/interfaces";
-import Big from "big.js";
 import { CurrencyExt } from "../types";
-import { isCurrencyEqual } from "../utils";
-
-// TODO: move type definitions to separate file later
-
-enum PoolType {
-    STANDARD = "STANDARD",
-    STABLE = "STABLE",
-}
-
-type LPToken = CurrencyExt; // TODO: specify when the currencies are refactored to have LP token type
-
-type PooledCurrencies = Array<MonetaryAmount<CurrencyExt>>;
-
-interface LiquidityPoolBase {
-    type: PoolType;
-    lpToken: LPToken;
-    pooledCurrencies: PooledCurrencies; // Array of 2 for standard pools, array of 2+ for stable pools.
-    apr: string; // Percentage.
-    tradingFee: string; // Percentage.
-}
-
-interface TradingPair {
-    type: PoolType;
-    token0: CurrencyExt;
-    token1: CurrencyExt;
-    reserve0: MonetaryAmount<CurrencyExt>;
-    reserve1: MonetaryAmount<CurrencyExt>;
-    // NOTE: do not throw here, rather return 0
-    getOutputAmount: (inputAmount: MonetaryAmount<CurrencyExt>) => [MonetaryAmount<CurrencyExt>, LiquidityPool];
-    pathOf: (inputCurrency: CurrencyExt) => MultiPathElement;
-}
-
-export interface MultiPathElement {
-    stable: boolean;
-    input: CurrencyExt;
-    output: CurrencyExt;
-    pair: TradingPair;
-    pool?: StableLiquidityPool;
-    basePool?: StableLiquidityPool;
-    fromBase?: boolean;
-}
-
-type MultiPath = Array<MultiPathElement>;
-
-class StandardLiquidityPool implements LiquidityPoolBase, TradingPair {
-    public type = PoolType.STANDARD;
-    public token0: CurrencyExt;
-    public token1: CurrencyExt;
-    public reserve0: MonetaryAmount<CurrencyExt>;
-    public reserve1: MonetaryAmount<CurrencyExt>;
-    constructor(
-        public lpToken: LPToken,
-        public pooledCurrencies: PooledCurrencies,
-        public apr: string,
-        public tradingFee: string
-    ) {
-        if (pooledCurrencies.length !== 2) {
-            throw new Error("Standard liquidity pool has to always consist of 2 currencies!");
-        }
-        this.token0 = pooledCurrencies[0].currency;
-        this.token1 = pooledCurrencies[1].currency;
-        this.reserve0 = pooledCurrencies[0];
-        this.reserve1 = pooledCurrencies[1];
-    }
-
-    public pathOf(inputCurrency: CurrencyExt): MultiPathElement {
-        return {
-            stable: false,
-            input: inputCurrency,
-            output: isCurrencyEqual(inputCurrency, this.token0) ? this.token1 : this.token0,
-            pair: this,
-        };
-    }
-
-    public getOutputAmount(
-        inputAmount: MonetaryAmount<CurrencyExt>
-    ): [MonetaryAmount<CurrencyExt>, StandardLiquidityPool] {
-        // TODO
-        return [] as any;
-    }
-}
-
-class StableLiquidityPool implements LiquidityPoolBase {
-    public type = PoolType.STABLE;
-    constructor(
-        public lpToken: LPToken,
-        public pooledCurrencies: PooledCurrencies,
-        public apr: string,
-        public tradingFee: string,
-        public poolId: number
-    ) {}
-}
-
-type LiquidityPool = StandardLiquidityPool | StableLiquidityPool;
-
-// TODO: improve, simplify, verify computation
-const computeMiddlePrice = (path: MultiPath, inputAmount: MonetaryAmount<CurrencyExt>): Big => {
-    const prices: Big[] = [];
-    const currencyPath = [inputAmount.currency, ...path.map((pathElement) => pathElement.output)];
-
-    let currentInputAmount = inputAmount;
-    for (const [i, pathElement] of path.entries()) {
-        let currentPrice: Big;
-        if (pathElement.stable) {
-            // TODO: how to compute middle price for curve pools?
-            //
-            // const outputAmount = getStableSwapOutputAmount(pathElement, currentInputAmount);
-            // currentPrice = new Price(
-            //     pathElement.input,
-            //     pathElement.output,
-            //     JSBI.multiply(ONE, currentInputAmount.decimalScale),
-            //     JSBI.multiply(ONE, outputAmount.decimalScale)
-            // );
-            // currentInputAmount = outputAmount;
-        } else {
-            const pair = pathElement.pair;
-            if (isCurrencyEqual(currencyPath[i], pair.token0)) {
-                // TODO: can we put this into 1 monetary amount variable and use only that ??
-                currentPrice = pair.reserve1.toBig().div(pair.reserve0.toBig());
-                currentInputAmount = new MonetaryAmount(pair.token1, currentInputAmount.mul(currentPrice).toBig());
-            } else {
-                currentPrice = pair.reserve0.toBig().div(pair.reserve1.toBig());
-                currentInputAmount = new MonetaryAmount(pair.token0, currentInputAmount.mul(currentPrice).toBig());
-            }
-        }
-
-        prices.push(currentPrice);
-    }
-
-    return prices.slice(1).reduce((accumulator, currentValue) => accumulator.multiply(currentValue), prices[0]);
-};
-
-const computePriceImpact = (
-    path: MultiPath,
-    inputAmount: MonetaryAmount<CurrencyExt>,
-    outputAmount: MonetaryAmount<CurrencyExt>
-): string => {
-    const midPrice = computeMiddlePrice(path, inputAmount);
-    const exactQuote = midPrice.mul(inputAmount.toBig());
-    // calculate priceImpact := (exactQuote - outputAmount) / exactQuote
-    const priceImpact = exactQuote.sub(outputAmount.toBig()).div(exactQuote);
-    // Return percentage.
-    return priceImpact.mul(100).toString();
-};
-
-class Trade {
-    public executionPrice: MonetaryAmount<CurrencyExt>;
-    public priceImpact: string; // Percentage.
-    constructor(
-        public path: MultiPath, // Is empty array if no path was found.
-        public inputAmount: MonetaryAmount<CurrencyExt>,
-        public outputAmount: MonetaryAmount<CurrencyExt>
-    ) {
-        const rawPrice = outputAmount.toBig().div(inputAmount.toBig());
-        this.executionPrice = new MonetaryAmount(outputAmount.currency, rawPrice);
-        this.priceImpact = computePriceImpact(path, inputAmount, outputAmount);
-    }
-
-    public isBetter(anotherTrade: Trade | null): boolean {
-        if (anotherTrade === null) {
-            return true;
-        }
-
-        if (
-            !this.inputAmount.eq(anotherTrade.inputAmount) ||
-            !isCurrencyEqual(this.outputAmount.currency, anotherTrade.outputAmount.currency)
-        ) {
-            throw new Error("Trade: isBetterThan: Comparing 2 different trades is not possible.");
-        }
-
-        // TODO: extend comparator in case of same output amount but different paths,
-        //       prefer trade with lower price impact
-        return this.outputAmount.gte(anotherTrade.outputAmount);
-    }
-
-    /**
-     * Get minimum output amount for trade with provided maximum allowed slippage.
-     *
-     * @param {number} maxSlippage Maximum slippage in percentage.
-     * @returns {MonetaryAmount<CurrencyExt>} Minimum output amount of trade allowed with provided slippage.
-     */
-    public getMinimumOutputAmount(maxSlippage: number): MonetaryAmount<CurrencyExt> {
-        const maxSlippageInDecimal = maxSlippage / 100;
-        const amount = Big(1).sub(maxSlippageInDecimal).mul(this.outputAmount.toBig());
-        const minimumAmountOut = new MonetaryAmount(this.outputAmount.currency, amount);
-
-        return minimumAmountOut;
-    }
-}
-
-const isStandardPool = (pool: LiquidityPoolBase): pool is StandardLiquidityPool => pool.type === PoolType.STANDARD;
-const isStablePool = (pool: LiquidityPoolBase): pool is StableLiquidityPool => pool.type === PoolType.STABLE;
-
-const convertStablePoolToTradingPairs = (pool: StableLiquidityPool): Array<TradingPair> => {
-    // TODO
-    return [];
-};
-
-/**
- * Get all trading pairs based on provided pools.
- *
- * @param pools All standard and stable pools.
- * @returns {Array<TradingPair>} All trading pairs.
- */
-const getAllTradingPairs = (pools: Array<LiquidityPool>): Array<TradingPair> => {
-    const pairs: Array<TradingPair> = [];
-    pools.forEach((pool) => {
-        if (isStandardPool(pool)) {
-            pairs.push(pool);
-        } else {
-            const stablePairs = convertStablePoolToTradingPairs(pool);
-            pairs.push(...stablePairs);
-        }
-    });
-
-    return pairs;
-};
+import { LiquidityPool } from "./amm/liquidity-pool/types";
+import { getAllTradingPairs } from "./amm/liquidity-pool/utils";
+import { Trade } from "./amm/trade/trade";
+import { LPToken, PooledCurrencies, PoolType } from "./amm/types";
+import { findBestTradeRecursively } from "./amm/utils";
 
 const HOP_LIMIT = 4; // TODO: add as parameter?
-
-const findBestTradeRecursively = (
-    inputAmount: MonetaryAmount<CurrencyExt>,
-    outputCurrency: CurrencyExt,
-    pairs: Array<TradingPair>,
-    hopLimit: number,
-    path: MultiPath,
-    originalInputAmount = inputAmount
-): Trade | null => {
-    if (hopLimit <= 0) {
-        return null;
-    }
-    const inputCurrency = inputAmount.currency;
-    let bestTrade: Trade | null = null;
-
-    for (const pair of pairs) {
-        if (isCurrencyEqual(inputCurrency, pair.token0) || isCurrencyEqual(inputCurrency, pair.token1)) {
-            // Skip iteration if input currency is not part of current pair.
-            continue;
-        }
-        const [outputAmount] = pair.getOutputAmount(inputAmount);
-        const currentPath = [...path, pair.pathOf(inputCurrency)];
-
-        // Complete path is found
-        if (isCurrencyEqual(outputCurrency, outputAmount.currency)) {
-            const trade = new Trade(currentPath, originalInputAmount, outputAmount);
-            if (trade.isBetter(bestTrade)) {
-                bestTrade = trade;
-            }
-        } else {
-            // Recursively check starting from current pair with decreased hop amount.
-            const trade = findBestTradeRecursively(
-                outputAmount,
-                outputCurrency,
-                pairs,
-                hopLimit - 1,
-                currentPath,
-                originalInputAmount
-            );
-            if (trade !== null && trade.isBetter(bestTrade)) {
-                bestTrade = trade;
-            }
-        }
-    }
-
-    return bestTrade;
-};
 
 export interface AMMAPI {
     /**
@@ -328,9 +68,9 @@ export interface AMMAPI {
     /**
      * Get all liquidity pools.
      *
-     * @returns {Promise<Array<LiquidityPoolBase>>} All liquidity pools.
+     * @returns {Promise<Array<LiquidityPool>>} All liquidity pools.
      */
-    getLiquidityPools(): Promise<Array<LiquidityPoolBase>>;
+    getLiquidityPools(): Promise<Array<LiquidityPool>>;
 
     /**
      * Swap assets.
@@ -359,11 +99,16 @@ export interface AMMAPI {
      * Removes liquidity from pool.
      *
      * @param {MonetaryAmount<LPToken>} amount Amount of LP token to be removed
+     * @param {LiquidityPool} pool Liquidity pool to remove from.
      * @param customCurrenciesProportion Optional parameter that allows to specify proportion
      *        of pooled currencies in which the liquidity should be withdrawn.
      * @note Removes `amount` of liquidity in LP token, breaks it down and transfers to account.
      */
-    removeLiquidity(amount: MonetaryAmount<LPToken>, customCurrenciesProportion?: PooledCurrencies): Promise<void>;
+    removeLiquidity(
+        amount: MonetaryAmount<LPToken>,
+        pool: LiquidityPool,
+        customCurrenciesProportion?: PooledCurrencies
+    ): Promise<void>;
 }
 
 export class DefaultAMMAPI implements AMMAPI {
@@ -380,7 +125,7 @@ export class DefaultAMMAPI implements AMMAPI {
             return null;
         }
 
-        return findBestTradeRecursively(inputAmount, outputCurrency, pairs, HOP_LIMIT, []);
+        return findBestTradeRecursively(inputAmount, outputCurrency, pairs, HOP_LIMIT);
     }
 
     getExpectedLiquidityDepositAmounts(
@@ -405,7 +150,7 @@ export class DefaultAMMAPI implements AMMAPI {
     getLiquidityProvidedByAccount(accountId: AccountId): Promise<MonetaryAmount<CurrencyExt>[]> {
         throw new Error("Method not implemented.");
     }
-    getLiquidityPools(): Promise<Array<LiquidityPoolBase>> {
+    getLiquidityPools(): Promise<Array<LiquidityPool>> {
         throw new Error("Method not implemented.");
     }
     swap(
@@ -423,7 +168,11 @@ export class DefaultAMMAPI implements AMMAPI {
             //
         });
     }
-    removeLiquidity(amount: MonetaryAmount<LPToken>, customCurrenciesProportion?: PooledCurrencies): Promise<void> {
+    removeLiquidity(
+        amount: MonetaryAmount<LPToken>,
+        pool: LiquidityPool,
+        customCurrenciesProportion?: PooledCurrencies
+    ): Promise<void> {
         return new Promise(function () {
             //
         });
